feat: add ci, minor test adjustments and use toml only
