+++ .github/workflows/ci.yml
@@ -1,0 +1,21 @@
+name: test
+on:
+  push:
+    branches:
+      - '**'
+  pull_request:
+    branches:
+      - '**'
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    container: python:3.8-slim
+    steps:
+      - uses: actions/checkout@v2
+      - name: Install prerequisites (for OpenCV)
+        run: apt-get update && apt-get install ffmpeg libsm6 libxext6 -y
+      - name: Install trajdata base version
+        run: python -m pip install .
+      - name: Run tests
+        run: python -m unittest tests/test_state.py
