+++ pyproject.toml
@@ -2,8 +2,47 @@
 profile = "black"
 
 [build-system]
-requires = [
-    "setuptools>=58",
-    "wheel"
-]
+requires = ["setuptools>=58", "wheel"]
 build-backend = "setuptools.build_meta"
+
+[project]
+classifiers = [
+    "Development Status :: 3 - Alpha",
+    "Intended Audience :: Developers",
+    "Programming Language :: Python :: 3.8",
+    "License :: OSI Approved :: Apache Software License",
+]
+name = "trajdata"
+version = "1.3.2"
+authors = [{ name = "Boris Ivanovic", email = "bivanovic@nvidia.com" }]
+description = "A unified interface to many trajectory forecasting datasets."
+readme = "README.md"
+requires-python = ">=3.8"
+dependencies = [
+    "numpy>=1.19",
+    "tqdm>=4.62",
+    "matplotlib>=3.5",
+    "dill>=0.3.4",
+    "pandas>=1.4.1",
+    "pyarrow>=7.0.0",
+    "torch>=1.10.2",
+    "zarr>=2.11.0",
+    "kornia>=0.6.4",
+    "seaborn>=0.12",
+    "bokeh>=3.0.3",
+    "geopandas>=0.13.2",
+    "protobuf==3.19.4",
+    "scipy>=1.9.0",
+    "opencv-python>=4.5.0"
+]
+
+[project.optional-dependencies]
+dev = ["black", "isort", "pytest", "pytest-xdist", "twine", "build"]
+interaction = ["lanelet2==1.2.1"]
+lyft = ["l5kit==1.5.0"]
+nusc = ["nuscenes-devkit==1.1.9"]
+waymo = ["tensorflow==2.11.0", "waymo-open-dataset-tf-2-11-0", "intervaltree"]
+
+[project.urls]
+"Homepage" = "https://github.com/nvr-avg/trajdata"
+"Bug Tracker" = "https://github.com/nvr-avg/trajdata/issues"
