--- requirements.txt
@@ -1,31 +1,0 @@
-numpy
-tqdm
-matplotlib
-dill
-pandas
-seaborn
-pyarrow
-torch
-zarr
-kornia
-bokeh
-
-# nuScenes devkit
-nuscenes-devkit==1.1.9
-
-# Lyft Level 5 devkit
-protobuf==3.19.4
-l5kit==1.5.0
-
-# Waymo devkit
-tensorflow==2.11.0
-waymo-open-dataset-tf-2-11-0
-intervaltree
-
-# Development
-black
-isort
-pytest
-pytest-xdist
-twine
-build
