--- setup.cfg
@@ -1,56 +1,0 @@
-[metadata]
-name = trajdata
-version = 1.3.2
-author = Boris Ivanovic
-author_email = bivanovic@nvidia.com
-description = A unified interface to many trajectory forecasting datasets.
-long_description = file: README.md
-long_description_content_type = text/markdown
-license = Apache License 2.0
-url = https://github.com/nvr-avg/trajdata
-classifiers =
-    Development Status :: 3 - Alpha
-    Intended Audience :: Developers
-    Programming Language :: Python :: 3.8
-    License :: OSI Approved :: Apache Software License
-
-[options]
-package_dir =
-    = src
-packages = find:
-python_requires = >=3.8
-install_requires =
-    numpy>=1.19
-    tqdm>=4.62
-    matplotlib>=3.5
-    dill>=0.3.4
-    pandas>=1.4.1
-    pyarrow>=7.0.0
-    torch>=1.10.2
-    zarr>=2.11.0
-    kornia>=0.6.4
-    seaborn>=0.12
-    bokeh>=3.0.3
-
-[options.packages.find]
-where = src
-
-[options.extras_require]
-dev =
-    black
-    isort
-    pytest
-    pytest-xdist
-    twine
-    build
-nusc =
-    nuscenes-devkit==1.1.9
-lyft =
-    protobuf==3.19.4
-    l5kit==1.5.0
-waymo =
-    tensorflow==2.11.0
-    waymo-open-dataset-tf-2-11-0
-    intervaltree
-interaction =
-    lanelet2==1.2.1
